Add screen event to system events
