+++ src/Schema/Events/System.ts
@@ -55,3 +55,27 @@ export interface SaleScreenLoadComplete {
   load_time_ms: number
   number_of_lots: number
 }
+
+/**
+ * A user views a screen on iOS
+ *
+ * This schema describes events sent to Segment from [[screen]]
+ *
+ *  @example
+ *  ```
+ *  {
+ *    action: "screen",
+ *    context_screen_owner_type: "artist",
+ *    context_screen_owner_id: "527ac4a0cd530e258d0000d0",
+ *    context_screen_owner_slug: "ramiro-gomez",
+ *    context_screen_referrer_type: "artwork"
+ *  }
+ * ```
+ */
+export interface Screen {
+  action: ActionType.screen
+  context_screen_owner_type: ScreenOwnerType
+  context_screen_owner_id?: string
+  context_screen_owner_slug?: string
+  context_screen_referrer_type?: ScreenOwnerType
+}
